Fix _create_t method for more recent releases of Module::Starter::Simple
In version 1.70 of Module::Starter::Simple a new parameter was added to
the _create_t method that allows to specify in which directory a test
script is supposed to go.  This change breaks the creation of new
distributions in Module::Starter::Smart with its own _create_t method,
because the parameter was added before the already existing parameters.
As a result every newly created test is written to a file with the name
of the directory it was supposed to be created in, containing only the
filename of the test instead of the actual test script.

Fix this by making the replacement _create_t method accept its arguments
either via the old or the augmented argument list, defaulting to the ./t
directory if no test directory was specified. Also add a test for
_create_t to check both the old and new behavior.